+++ src/WebScraper/ActressJavDatabase.cs
@@ -3,6 +3,7 @@ using AngleSharp.Html.Dom;
 using Common;
 using MovieInfo;
 using System;
+using System.Runtime.Remoting.Contexts;
 
 namespace WebScraper
 {
@@ -53,75 +54,56 @@ namespace WebScraper
                         }
                     }               
                 }
+                if (element.NodeName == "H1")
+                {
+                    if (element.TextContent == Actress.Name)
+                    {
+                        if (element.Parent == null)
+                            continue;
+                        string content = element.Parent.TextContent;
+                        if (String.IsNullOrEmpty(content))
+                            continue;
 
-                if (element.TextContent == "Japanese Name")
-                {
-                    var nextSibling = element.NextSibling?.NextSibling;
-                    if (IsValidNode(nextSibling))
-                    {
-                        Actress.JapaneseName = nextSibling.TextContent.Trim();
-                    }
-                }
-                else if (element.TextContent == "Date of Birth")
-                {
-                    var nextSibling = element.NextSibling?.NextSibling;
-                    if (IsValidNode(nextSibling))
-                    {
-                        string[] dateParts = nextSibling.TextContent.Split('-');
+                        // Parse Japanese name
+                        Actress.JapaneseName = Parse(content, "JP: ");
+
+                        // Parse date of birth
+                        string dobText = Parse(content, "DOB: ");
                         int year = 0;
                         int month = 0;
                         int day = 0;
-                        Utilities.StringToDateTime(nextSibling.TextContent, out year, out month, out day);
+                        Utilities.StringToDateTime(dobText, out year, out month, out day);
                         Actress.DobYear = year;
                         Actress.DobMonth = month;
                         Actress.DobDay = day;
-                    }
-                }
-                else if (element.TextContent == "Height")
-                {
-                    var nextSibling = element.NextSibling?.NextSibling;
-                    if (IsValidNode(nextSibling))
-                    {
-                        Actress.Height = Utilities.ParseInitialDigits(nextSibling.TextContent);
-                    }
-                }
-                else if (element.TextContent == "Cup")
-                {
-                    var nextSibling = element.NextSibling?.NextSibling;
-                    if (IsValidNode(nextSibling))
-                    {
-                        string cupText = nextSibling.TextContent.Trim();
-                        // Saw one example of multiple entries.
-                        string[] cups = cupText.Split(' ');
-                        Actress.Cup = cups[0];
-                    }
-                }
-                else if (element.TextContent == "Measurements")
-                {
-                    var nextSibling = element.NextSibling?.NextSibling;
-                    if (IsValidNode(nextSibling))
-                    {
-                        string[] dateParts = nextSibling.TextContent.Split('-');
+
+                        // Parse height
+                        string height = Parse(content, "Height: ");
+                        Actress.Height = Utilities.ParseInitialDigits(height);
+
+                        // Parse measurements
+                        string measurements = Parse(content, "Measurements: ");
+                        string[] dateParts = measurements.Split('-');
                         try
                         {
                             Actress.Bust = int.Parse(dateParts[0]);
                             Actress.Waist = int.Parse(dateParts[1]);
                             Actress.Hips = int.Parse(dateParts[2]);
                         }
-                        catch (Exception)
+                        catch (Exception) 
                         { }
-                    }
-                }
-                else if (element.TextContent == "Blood Type")
-                {
-                    var nextSibling = element.NextSibling?.NextSibling;
-                    if (IsValidNode(nextSibling))
-                    {
-                        Actress.BloodType = nextSibling.TextContent.Trim();
+
+                        // Parse cup
+                        string cupText = Parse(content, "Cup: ");
+                        // Saw one example of multiple entries.
+                        string[] cups = cupText.Split(' ');
+                        Actress.Cup = cups[0];
+
+                        // Parse blood type
+                        Actress.BloodType = Parse(content, "Blood: ");
                     }
                 }
             }
-
         }
 
         protected override bool IsLanguageSupported()
@@ -135,6 +117,22 @@ namespace WebScraper
 
         #region Private Functions
 
+        private string Parse(string text, string search)
+        {
+            string retVal = String.Empty;
+            int index = text.IndexOf(search, StringComparison.Ordinal);
+            if (index == -1)
+                return retVal;
+            string substr = text.Substring(index + search.Length);
+            if (substr.Length == 0)
+                return retVal;
+            int wsIndex = substr.IndexOf(' ');
+            int nlIndex = substr.IndexOf('\n');
+            index = Math.Min(wsIndex, nlIndex);
+            retVal = substr.Substring(0, index);
+            return retVal;
+        }
+        /*
         private bool IsValidNode(INode element)
         {
             if (element == null)
@@ -145,7 +143,7 @@ namespace WebScraper
                 return false;
             return true;
         }
-
+        */
         #endregion
     }
 }